WIP pytorch optimization
