+++ RWKV-v5/src/module/CoreDependencies.py
@@ -1,0 +1,127 @@
+### ---
+# External shared pytorch dependencies, used across multiple modules
+#
+# Other modules in this codebase should not be imported in this file, 
+# only external dependencies, and shared dependencies should be imported here
+### ---
+
+import gc, math, os
+from random import randint
+import time
+from typing import List, Optional
+
+import numpy as np
+import torch
+# torch._C._jit_set_profiling_executor(True)
+# torch._C._jit_set_profiling_mode(True)
+import torch.nn as nn
+from torch.nn import functional as F
+
+import lightning as L
+from lightning.pytorch.utilities import rank_zero_info, rank_zero_only
+from lightning.pytorch.strategies import DeepSpeedStrategy
+
+import deepspeed
+from deepspeed.ops.adam import DeepSpeedCPUAdam, FusedAdam
+import deepspeed.runtime.lr_schedules
+import wandb
+
+from torch.utils.cpp_extension import load
+
+# Script dir for various files
+SCRIPT_DIR = os.path.dirname(os.path.realpath(__file__))
+CUDA_DIR = os.path.abspath(os.path.join(SCRIPT_DIR, "../cuda"))
+
+### ---
+# JIT / torch compile special handling
+### ---
+
+# Because we are using APIs avaliable only to pytorch 2.1.0
+# We will throw an error if the user is using a lower version
+from packaging import version
+def is_torch_version_above(required_version):
+    torch_version = version.parse(torch.__version__.split('+')[0])
+    return torch_version >= version.parse(required_version)
+IS_TORCH_2_1_COMPATIBLE = is_torch_version_above("2.0.9999")
+
+# Get the JIT / torch compile option flags from the environment
+# This default is FOR inference mode, the trainer mode default is configured in the lightning_trainer.py
+RWKV_JIT_ON         = os.getenv("RWKV_JIT_ON", "1").lower() in ("1", "true", "yes")
+RWKV_TORCH_COMPILE  = os.getenv("RWKV_TORCH_COMPILE", f"0").lower() in ("1", "true", "yes")
+RWKV_TORCH_RUN_MODE = None
+
+# Disable torch compile if its not atleast v2.1
+if not IS_TORCH_2_1_COMPATIBLE:
+    RWKV_TORCH_COMPILE = False
+
+# We enable JITMod*/Function when supporting torch.jit
+# We use TorchCompile* when supporting torch compile
+# based on the current runtime settings
+if RWKV_TORCH_COMPILE:
+    RWKV_TORCH_RUN_MODE = "torch-compile"
+
+    JITModClass  = nn.Module
+    JITModMethod = lambda x: x
+    JITFunction  = lambda x: x
+
+    # PS: i have tried mode="max-autotune", and mode="reduce-overhead", however they crash
+    #     for now (8th July 2023). I may introduce them in the future once they are stable
+    #
+    #     Additionally, torch.compile has issues with the pytorch.lightning module directly
+    # ---
+
+    # We generally have 2 major options, either we use torch.compile
+    # onto the key top level functions (train, val, test, predict, etc)
+    # and let the compiler handle all the decision making on how to optimize
+    #
+    # However this was found to basically just match JIT level of performance exactly
+    # ---
+    # TCompileMax          = lambda x: x
+    # TCompileBaseline     = lambda x: torch.compile(x, fullgraph=False)
+
+    # Alternatively, we can perform a much more aggressive optimization on critical functions
+    # that we know are compatible with torch.compile(fullgraph=True) - which provides the highest
+    # level of optimization possible with torch.compile
+    # ---
+    TCompileMax        = lambda x: torch.compile(x, mode="max-autotune", fullgraph=True)
+    TCompileBaseline   = lambda x: x
+
+    # ---
+    # Because torch.compile is expected to change overtime, the two options should 
+    # be tested every now and then, for any performance changes
+    #
+    # and we should switch over to the broaded automated approach if its "faster"
+    # ---
+
+    # Used to wrap functions which are **not** torch.compile compatible
+    TCompileDisable    = torch._dynamo.disable
+
+    # The following are known warnings in the nightly build, that can be safely ignored for stable release
+    #
+    # `torch._inductor.utils: [WARNING] DeviceCopy in input program` 
+    # https://discuss.pytorch.org/t/what-can-cause-warning-devicecopy-in-input-program/175566
+
+elif RWKV_JIT_ON:
+    RWKV_TORCH_RUN_MODE = "torch-jit"
+    JITModClass  = torch.jit.ScriptModule
+    JITModMethod = torch.jit.script_method
+    JITFunction  = torch.jit.script
+
+    # JITModClass  = nn.Module
+    # JITModMethod = lambda x: x
+    # JITFunction  = lambda x: x
+
+    TCompileMax        = lambda x: x
+    TCompileBaseline   = lambda x: x
+    TCompileDisable    = lambda x: x
+else:
+    RWKV_TORCH_RUN_MODE = "torch-native"
+    JITModClass  = nn.Module
+    JITModMethod = lambda x: x
+    JITFunction  = lambda x: x
+
+    TCompileMax        = lambda x: x
+    TCompileBaseline   = lambda x: x
+    TCompileDisable    = lambda x: x
+
+print(f"[RWKV.model] Running RWKV model using '{RWKV_TORCH_RUN_MODE}' with torch '{torch.__version__}'")
